webmidi/midioutputmap.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="UTF-8">
    <title>obtain & list MIDI output ports | Web MIDI API Test</title>
    <link rel="author" title="Ryoya KAWAI" href="mailto:ryoya.kawai@gmail.com">
    <link rel="help" href="http://webaudio.github.io/web-midi-api/#midioutputmap-interface"> <!-- 4.4 -->
    <link rel="help" href="http://webaudio.github.io/web-midi-api/#midiport-interface"> <!-- 6. -->
    <meta name="assert" content="Return true when test success in outputs().">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <p>Success when be obtained OUTPUT MIDI devce and its properties are correct.</p>
    <div id="warning" style="color:red;"></div>
    <div id="log"></div>
    <!-- PAGE CONTENT -->
    <script type="text/javascript">
    var MIDI;
    var test_midiOutputMap = async_test("MIDIOutputMap Interface Test");
    navigator.requestMIDIAccess().then( successCallback, errorCallback );
    function successCallback(access) {
        MIDI = access;
        test_midiOutputMap.step(function() {
            var outputs=[];
            var outputIterator=MIDI.outputs.values();
            for(var o=outputIterator.next(); !o.done; o=outputIterator.next()) {
                outputs.push(o.value);
            }
            if(outputs.length==0) {
                document.getElementById("warning").innerHTML="Your Computer looks no <b><u>MIDI OUTPUT</u></b> devices are pluged-in.<br> Please plug-in at least one <b><u>MIDI OUTPUT</u></b> device before run this Test.";
                assert_unreached("No MIDI OUTPUT devices are pluged-in");
                test_midiOutputMap.done();
            } else {
                // 4.4 MIDIOutputMap Interface
                test(function() {
                    assert_class_string(MIDI.outputs, "MIDIOutputMap");
                }, "4.4 MIDIOutputMap Interface Constructor name Test");
                // 4.4.2 Attributes Test
                test(function() {
                    assert_equals("number", typeof MIDI.outputs.size);
                }, "4.4.2 Attributes Test (MIDIOutputMap interface)");
                // 4.4.3 Methods Test
                test(function() {
                    assert_equals("function", typeof MIDI.outputs.entries);
                    var outputEntries=MIDI.outputs.entries()
                    assert_class_string(outputEntries, "Iterator");
                }, "4.4.3 Methods Test (MIDIOutputMap interface) : entries()");
                /*
                test(function() {
                    assert_equals("function", typeof MIDI.outputs.forEach);
                    var outputEntries=MIDI.outputs.forEach()
                    assert_class_string(outputEntries, "undefined");
                }, "4.4.3 Methods Test (MIDIOutputMap interface) : forEach()");
                 */
                test(function() {
                    assert_equals("function", typeof MIDI.outputs.keys);
                    var outputEntries=MIDI.outputs.keys();
                    assert_class_string(outputEntries, "Iterator");
                }, "4.4.3 Methods Test (MIDIOutputMap interface) : keys()");
                test(function() {
                    assert_equals("function", typeof MIDI.outputs.get);
                    var keyIterator=MIDI.outputs.keys();
                    for(var itr=keyIterator.next(); !itr.done; itr=keyIterator.next()) {
                        var outputEntries=MIDI.outputs.get(itr.value);
                        assert_class_string(outputEntries, "MIDIOutput");
                    }
                }, "4.4.3 Methods Test (MIDIOutputMap interface) : get()");
                test(function() {
                    assert_equals("function", typeof MIDI.outputs.has);
                    var keyIterator=MIDI.outputs.keys();
                    for(var itr=keyIterator.next(); !itr.done; itr=keyIterator.next()) {
                        var outputEntries=MIDI.outputs.has(itr.value);
                        assert_class_string(outputEntries, "Boolean");
                    }
                }, "4.4.3 Methods Test (MIDIOutputMap interface) : has()");
                test(function() {
                    assert_equals("function", typeof MIDI.outputs.values);
                    var outputEntries=MIDI.outputs.values();
                    assert_class_string(outputEntries, "Iterator");
                }, "4.4.3 Methods Test (MIDIOutputMap interface) : values()");
                for( var j=0; j<outputs.length; j++ ) {
                    // 6. MIDIPort Interface (output)
                    test(function() {
                        assert_idl_attribute(outputs[j], "id");
                        assert_idl_attribute(outputs[j], "manufacturer");
                        assert_idl_attribute(outputs[j], "name");
                        assert_idl_attribute(outputs[j], "type");
                        assert_idl_attribute(outputs[j], "version");
                        assert_idl_attribute(outputs[j], "state");
                        assert_idl_attribute(outputs[j], "connection");
                        assert_idl_attribute(outputs[j], "onstatechange");
                    }, "6.1 MIDIPort Interface Attributes Name Test (output)" + " (#"+j+")");

                    test(function() {
                        assert_readonly(outputs[j], "id");
                        assert_readonly(outputs[j], "manufacturer");
                        assert_readonly(outputs[j], "name");
                        assert_readonly(outputs[j], "type");
                        assert_readonly(outputs[j], "version");
                        assert_readonly(outputs[j], "state");
                        assert_readonly(outputs[j], "connection");

                        assert_equals("string", typeof outputs[j]["id"], "[Type of \"id\": String] " );
                        assert_true( typeof outputs[j]["manufacturer"]=="string" || outputs[j]["manufacturer"]==null, "[Type of \"manufacturer\": String or null] " );
                        assert_true( typeof outputs[j]["name"]=="string" || outputs[j]["name"]==null, "[Type of \"name\": String or null] " );
                        assert_in_array( outputs[j]["type"], ["input", "output"], "[Type of \"type\": input or output] " );
                        assert_true( typeof outputs[j]["version"]=="string" || outputs[j]["version"]==null, "[Type of \"version\": String or null] " );
                        assert_equals("object", typeof outputs[j]["onstatechange"]);
                        assert_class_string(outputs[j], "MIDIOutput");
                    }, "6.1 MIDIPort Interface Attributes Type Test (output)" + " (#"+j+")");

                    // 6.3 MIDIOutput Interface
                    test(function() {
                        assert_equals("function", typeof outputs[j].send);
                    }, "6.3.1 MIDIOutput Methods Type Test" + " (#"+j+")");
                    test(function() {
                        assert_idl_attribute(outputs[j], "send");
                    }, "6.3.1 MIDIOutput Methods Name Test" + " (#"+j+")");
                }
                test_midiOutputMap.done();
            }
        });
    }
    function errorCallback(message) {
        test_midiOutputMap.step(function() {
            assert_unreached("Called errorCallback in Promise of requestMIDIAccess");
            test_midiOutputMap.done();
        });
    }
    </script>
  </body>
</html>
